refactor(broadcast-app-indexer): move routes under /app
