+++ apps/broadcast-app-indexer/src/api/routes.ts
@@ -1,5 +1,5 @@
 import type { OpenAPIHono } from "@hono/zod-openapi";
-import { initializeChannelsRoutes } from "./channels/routes";
+import { initializeChannelsRoutes } from "./apps/[appId]/channels/routes";
 
 export async function initializeRoutes(api: OpenAPIHono): Promise<void> {
   await initializeChannelsRoutes(api);
